requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
dj-database-url==1.2.0
Django==4.0.4
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-debug-toolbar==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.26.2
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
generics==6.0.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
instaloader==4.9.6
jsonschema==4.17.3
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.2
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
